+++ tests/testthat/test-getAnchors.R
@@ -4,7 +4,8 @@ test_that("getAnchors works", {
     skip_on_cran() # don't run tests that depend on external resource on CRAN
     if (!is.null(labbcatCredentials(labbcat.url, "demo", "demo"))) skip("Server not available")
 
-    orthography <- getAnnotations(labbcat.url, "UC427_ViktoriaPapp_A_ENG.eaf", "orthography", 20, 0)
+    orthography <- getAnnotations(
+        labbcat.url, "UC427_ViktoriaPapp_A_ENG.eaf", "orthography", NULL, 20, 0)
     anchors <- getAnchors(labbcat.url, "UC427_ViktoriaPapp_A_ENG.eaf", orthography$startId)
     expect_equal(length(anchors$id), 20)
     expect_equal(length(anchors$offset), 20)